renamed latitude calculation file
